Move BadChange-type classes to svn_hook
Allows post-commit hook to raise exactly same error as pre-commit hook 
for syntax error in info file
