+++ lib/python/rosie/db_create.py
@@ -30,6 +30,7 @@ from rose.reporter import Reporter, Event
 from rose.resource import ResourceLocator
 from rosie.db import (
     LATEST_TABLE_NAME, MAIN_TABLE_NAME, META_TABLE_NAME, OPTIONAL_TABLE_NAME)
+from rosie.svn_hook import InfoFileError
 from rosie.svn_post_commit import RosieSvnPostCommitHook
 
 from rose.config import ConfigSyntaxError
@@ -203,7 +204,7 @@ class RosieDatabaseInitiator(object):
             try:
                 self.post_commit_hook.run(
                     repos_path, str(revision), no_notification=True)
-            except (ConfigSyntaxError, DatabaseError) as err:
+            except (ConfigSyntaxError, DatabaseError, InfoFileError) as err:
                 if sys.stdout.isatty():
                     sys.stdout.write("\r")
                     sys.stdout.flush()
